+++ src/__tests__/componentHelpers/FilterYears.tsx
@@ -10,6 +10,7 @@ export const FilterYearsTests = (
   endYear?: Filter,
   mockRouter?: MemoryRouter
 ) => {
+  const user = userEvent.setup({ delay: null });
   test("Renders Filter Years", () => {
     expect(
       screen.getByRole("spinbutton", {
@@ -27,7 +28,7 @@ export const FilterYearsTests = (
   });
 
   if (hasApplyButton) {
-    test("Submits filters with only 'from' value", () => {
+    test("Submits filters with only 'from' value", async () => {
       const yearGroup = screen.getByRole("group", {
         name: "Publication Year",
       });
@@ -39,7 +40,7 @@ export const FilterYearsTests = (
       });
       fireEvent.change(fromInput, { target: { value: 1990 } });
       expect(fromInput).toHaveValue(1990);
-      userEvent.click(applyButton);
+      await user.click(applyButton);
 
       expect(mockRouter).toMatchObject({
         pathname: "/search",
@@ -49,7 +50,7 @@ export const FilterYearsTests = (
         },
       });
     });
-    test("Submits filters with only 'to' value", () => {
+    test("Submits filters with only 'to' value", async () => {
       const yearGroup = screen.getByRole("group", {
         name: "Publication Year",
       });
@@ -60,7 +61,7 @@ export const FilterYearsTests = (
         name: "Apply",
       });
       fireEvent.change(toInput, { target: { value: 1990 } });
-      userEvent.click(applyButton);
+      fireEvent.click(applyButton);
 
       expect(mockRouter).toMatchObject({
         pathname: "/search",
@@ -71,7 +72,7 @@ export const FilterYearsTests = (
       });
     });
 
-    test("Submits search with both 'from' and 'to'", () => {
+    test("Submits search with both 'from' and 'to'", async () => {
       const yearGroup = screen.getByRole("group", {
         name: "Publication Year",
       });
@@ -86,7 +87,7 @@ export const FilterYearsTests = (
       });
       fireEvent.change(fromInput, { target: { value: 1990 } });
       fireEvent.change(toInput, { target: { value: 2000 } });
-      userEvent.click(applyButton);
+      await user.click(applyButton);
 
       expect(mockRouter).toMatchObject({
         pathname: "/search",
@@ -97,7 +98,7 @@ export const FilterYearsTests = (
       });
     });
 
-    test("shows error text when 'to' is after 'from", () => {
+    test("shows error text when 'to' is after 'from", async () => {
       const yearGroup = screen.getByRole("group", {
         name: "Publication Year",
       });
@@ -112,7 +113,7 @@ export const FilterYearsTests = (
       });
       fireEvent.change(fromInput, { target: { value: 1990 } });
       fireEvent.change(toInput, { target: { value: 1890 } });
-      userEvent.click(applyButton);
+      await user.click(applyButton);
 
       expect(
         screen.getByText("Start date must be before End date")